+++ src/hooks/use-toast/index.ts
@@ -8,6 +8,7 @@ type ToasterToast = ToastProps & {
   id: string
   title?: React.ReactNode
   description?: React.ReactNode
+  detail?: React.ReactNode
   action?: ToastActionElement
 }