+++ src/utils/api.ts
@@ -3,6 +3,8 @@ import { DataContextType } from './data-context'
 import { MessageEnum } from './message'
 import dbManager from './indexed-db'
 import { queryAndSendMessage } from './tab'
+import { AIError } from './error'
+import { sleep } from './promise'
 
 type BResponse<T> = {
   code: number
@@ -202,7 +204,7 @@ const connectAndStream = (message: { type: MessageEnum; data: any }) => {
             port.disconnect()
             break
           case 'error':
-            controller.error(new Error(response.error))
+            controller.error(new AIError(response.error, response.detail))
             port.disconnect()
             break
           case 'aborted':
@@ -215,7 +217,7 @@ const connectAndStream = (message: { type: MessageEnum; data: any }) => {
       port.onDisconnect.addListener(() => {
         const lastError = chrome.runtime.lastError
         if (lastError && !isCancelled) {
-          controller.error(new Error(lastError.message))
+          controller.error(new AIError('连接已断开', lastError.message))
         }
       })
 
@@ -231,11 +233,12 @@ const connectAndStream = (message: { type: MessageEnum; data: any }) => {
   }
 }
 
-const fetchChatGpt = async (titleArray: string[], config: AIConfig) => {
+const fetchChatGpt = async (titleArray: string[], config: AIConfig, useCustomAI: boolean) => {
   return connectAndStream({
     type: MessageEnum.fetchChatGpt,
     data: {
       titleArray,
+      useCustomAI,
       config: {
         apiKey: config.apiKey,
         baseURL: config.baseURL,
@@ -246,11 +249,17 @@ const fetchChatGpt = async (titleArray: string[], config: AIConfig) => {
   })
 }
 
-const fetchAIMove = async (videos: AIMoveInput, favoriteTitles: string[], config: AIMoveConfig) => {
+const fetchAIMove = async (
+  videos: AIMoveInput,
+  favoriteTitles: string[],
+  config: AIMoveConfig,
+  useCustomAI: boolean,
+) => {
   return connectAndStream({
     type: MessageEnum.fetchAIMove,
     data: {
       videos,
+      useCustomAI,
       favoriteTitles,
       config: {
         apiKey: config.apiKey,
@@ -295,8 +304,8 @@ const fetchAllFavoriteMedias = async (
   const mediaData = await dbManager.get(key)
   const isExpired = await dbManager.isExpired(key, expireTime)
   if (mediaData && !isExpired) return mediaData.data
-
   while (hasMore) {
+    await sleep(1000) // 防止触发b站api风控
     const response = await queryAndSendMessage<GetFavoriteListRes>({
       type: MessageEnum.getFavoriteList,
       data: { mediaId, pn: currentPage, ps: pageSize },