+++ src/utils/error.ts
@@ -1,0 +1,11 @@
+class AIError extends Error {
+  constructor(
+    message: string,
+    public detail?: string,
+  ) {
+    super()
+    this.message = message
+  }
+}
+
+export { AIError }