+++ src/utils/log.ts
@@ -1,7 +1,0 @@
-const log = (...args: any[]) => {
-  if (import.meta.env.DEV) {
-    console.log(...args)
-  }
-}
-
-export { log }